scripts/page_generator/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
beautifulsoup4==4.12.3
boto3==1.34.34
httpx==0.26.0 